remove live editing logic
