--- live-editing/configs/core/IConfigGenerator.ts
@@ -1,7 +1,0 @@
-import { Config } from "./Config";
-
-export type importsType = { [key: string]: string };
-export interface IConfigGenerator {
-    generateConfigs(): Config[];
-    additionalImports?: importsType;
-}
