+++ live-editing/generators/ConfigGenerators.ts
@@ -1,3 +1,4 @@
+import { Config, IConfigGenerator } from "igniteui-live-editing";
 import { ActionStripConfigGenerator } from "../configs/ActionStripConfigGenerator";
 import { AutocompleteConfigGenerator } from "../configs/AutocompleteConfigGenerator";
 import { AvatarConfigGenerator } from "../configs/AvatarConfigGenerator";
