--- live-editing/generators/MetaDataGenerator.ts
@@ -1,16 +1,0 @@
-
-// tslint:disable:prefer-const
-import * as fs from "fs";
-import { Generator } from "./Generator";
-import { MetaData } from "./misc/MetaData";
-
-export class MetaDataGenerator extends Generator {
-    constructor() {
-        super();
-    }
-
-    public Generate() {
-        let metadata = new MetaData();
-        fs.writeFileSync(this.getAssetsSamplesDir() + "/meta.json", JSON.stringify(metadata));
-    }
-}
