--- live-editing/generators/misc/MetaData.ts
@@ -1,6 +1,0 @@
-export class MetaData {
-    public generationTimeStamp: number;
-    constructor() {
-        this.generationTimeStamp = new Date().getTime();
-    }
-}
