--- live-editing/generators/misc/SampleDefinitionFile.ts
@@ -1,11 +1,0 @@
-import { LiveEditingFile } from "./LiveEditingFile";
-
-export class SampleDefinitionFile {
-    public sampleFiles: LiveEditingFile[];
-    public sampleDependencies: string;
-
-    constructor(files: LiveEditingFile[], dependencies: string[]) {
-        this.sampleFiles = files;
-        this.sampleDependencies = JSON.stringify(dependencies);
-    }
-}
