--- live-editing/generators/misc/SharedAssetsFile.ts
@@ -1,10 +1,0 @@
-import { LiveEditingFile } from "./LiveEditingFile";
-
-export class SharedAssetsFile {
-    public files: LiveEditingFile[];
-    public devDependencies: object;
-    constructor(files: LiveEditingFile[], devDependencies: object) {
-        this.files = files;
-        this.devDependencies = devDependencies;
-    }
-}
