--- live-editing/templates/app.component.ts.template
@@ -1,8 +1,0 @@
-import { Component } from "@angular/core";
-
-@Component({
-    selector: "app-root",
-    styleUrls: ["./app.component.scss"],
-    templateUrl: "./app.component.html"
-})
-export class AppComponent {}
