--- live-editing/templates/app.module.ts.template
@@ -1,17 +1,0 @@
-import { NgModule } from "@angular/core";
-import { FormsModule } from "@angular/forms";
-import { BrowserModule } from "@angular/platform-browser";
-import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
-import { AppComponent } from "./app.component";{imports}
-
-{additionalAdjustments}
-
-@NgModule({
-  bootstrap: [AppComponent],
-  declarations: [{ngDeclarations}],
-  imports: [{ngImports}],
-  providers: [{ngProviders}],
-  entryComponents: [{ngEntryComponents}],
-  schemas: [{ngSchemas}]
-})
-export class AppModule {}
